Fixed owner drawn cell element and added a more realistic demo
The GetCell call for the owner drawn element reused cells without changing the underlying custom element. This ment that as the user scrolled the table old cells were placed at the bottom on the grid. I modified the Get cell code to propogate the newly inserted OwnerDrawnElement down to the OwnerDrawnCellView.

If you want to see the prefix behaviour try the demo without the fix. You will see the first row data pop up at the bottom (original demo only worked because the cell content was always the same)