Making the sample be a universal app
I noticed that the MainWindow.xib cause the app to only accept user
input in a 320 x 480 section of the top left of the iPad. I removed
MainWindow.xib and moved the AppDelegate class to a separate file as is
done in the most recent templates. Then to launch the app, I
programmatically created the window and navigation controller in the
app delegate.