Update MonoTouch.Dialog/Utilities/ImageLoader.cs
Fix: Make sure target filename is updated when dealing with requests queued on the request queue.
Bug description: If all download threads are in use, a request is added to the request queue. When one of the download threads is free, it will pop the request from the request queue and start download it, but the target filename is incorrectly set to the path to the first download ever initiated by the download thread.