Added optional explicit summary to RootElement.
An explicit sumary may be useful for an application that may show a
derivative value (determined by a business rule) on the root element
summary as opposed to a direct binding to a child element value.

In my case I want to show an appointment by

[Customer name      StartTime - ToTime  >]

which requires the summar value be explicitly given to the root element.